requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
websockets>=12.0
pydantic>=2.6.0
tree-sitter>=0.23.0,<0.26
tree-sitter-python>=0.23.0
tree-sitter-php>=0.23.0
networkx>=3.2.1
httpx>=0.27.0
jedi>=0.19.2
python-dotenv>=1.0.0
aiofiles>=23.2.1
python-multipart>=0.0.9
